# Projects/backend/src/routes/auth.routes.ts
import { Router, Request, Response } from 'express';
import { logger } from '../utils/logger';

const router = Router();

/**
 * @route   POST /api/auth/register
 * @desc    Register a new user (placeholder)
 * @access  Public
 */
router.post('/register', (req: Request, res: Response) => {
  res.status(501).json({
    success: false,
    message: 'User registration not implemented yet',
    timestamp: new Date().toISOString()
  });
});

/**
 * @route   POST /api/auth/login
 * @desc    Login user (placeholder)
 * @access  Public
 */
router.post('/login', (req: Request, res: Response) => {
  res.status(501).json({
    success: false,
    message: 'User login not implemented yet',
    timestamp: new Date().toISOString()
  });
});

/**
 * @route   POST /api/auth/logout
 * @desc    Logout user (placeholder)
 * @access  Public
 */
router.post('/logout', (req: Request, res: Response) => {
  res.status(501).json({
    success: false,
    message: 'User logout not implemented yet',
    timestamp: new Date().toISOString()
  });
});

/**
 * @route   GET /api/auth/profile
 * @desc    Get user profile (placeholder)
 * @access  Private
 */
router.get('/profile', (req: Request, res: Response) => {
  res.status(501).json({
    success: false,
    message: 'User profile not implemented yet',
    timestamp: new Date().toISOString()
  });
});

export default router;